pyproject.toml
[tool.poetry]
name = "trans"
version = "0.1.0"
description = ""
authors = ["Young King <yanckin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.34.0"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
httpx = {extras = ["socks"], version = "^0.27.0"}
ruff = "^0.4.4"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

ignore = [
    "E501",  # line too long, handled by black
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
